fix more build warnings on Windows using gsl::narrow_cast
On Windows/MSVC we get warnings for implicit casts from size_t to int (and some other potentially lossy casts). In my debug Windows build in Qt Creator, fixing those takes the overall warning count from 270 down to 140. Most of the remainder are in 3rd party code I don't want to touch, or different warnings that need further investigation.

Using narrow_cast<> from the ISO Guidelines Support Library (GSL), via gsl-lite, an offshoot of Microsoft's open-source header-only GSL implementation. Unlike the Microsoft version, this one works with C++11 (Microsoft's requires C++14).

narrow_cast<> better expresses the intent of these casts (namely, "we believe these are safe"). It also has the benefit of being searchable and distinguishes them from other uses of static_cast.

The GSL also has narrow<>, which checks that the cast is safe and throws or terminates if it is not; I didn't use this.

The GSL is related to the ISO's C++ Core Guidelines, see http://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines.

The Guideline Support Library puts a few other interesting tidbits at our disposal, see: http://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#S-gsl

+++ src/cpp/r/RUtil.cpp
@@ -1,7 +1,7 @@
 /*
  * RUtil.cpp
  *
- * Copyright (C) 2009-18 by RStudio, Inc.
+ * Copyright (C) 2009-19 by RStudio, Inc.
  *
  * Unless you have received this program directly from RStudio pursuant
  * to the terms of a commercial license agreement with RStudio, then
@@ -15,6 +15,8 @@
 
 
 #include <r/RUtil.hpp>
+
+#include <gsl/gsl>
 
 #include <boost/algorithm/string/replace.hpp>
 #include <boost/regex.hpp>
@@ -182,7 +184,7 @@ core::Error iconvstr(const std::string& value,
             ::Riconv_close(handle);
             Error error = systemError(R_ERRNO, ERROR_LOCATION);
             error.addProperty("str", value);
-            error.addProperty("len", value.length());
+            error.addProperty("len", gsl::narrow_cast<int>(value.length()));
             error.addProperty("from", from);
             error.addProperty("to", to);
             return error;
